+++ src/main/java/gift/product/admin/AdminController.java
@@ -1,4 +1,4 @@
-package gift.admin;
+package gift.product.admin;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
