+++ src/main/java/gift/product/controller/ProductController.java
@@ -1,8 +1,8 @@
-package gift.controller;
+package gift.product.controller;
 
-import gift.dto.ProductRequestDto;
-import gift.model.Product;
-import gift.service.ProductService;
+import gift.product.dto.ProductRequestDto;
+import gift.product.model.Product;
+import gift.product.service.ProductService;
 import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
