+++ src/main/java/gift/product/exception/GlobalExceptionHandler.java
@@ -1,10 +1,9 @@
-package gift.exception;
+package gift.product.exception;
 
 import java.util.HashMap;
 import java.util.Map;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.ErrorResponse;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
