+++ src/main/java/gift/product/exception/ProductValidationException.java
@@ -1,4 +1,4 @@
-package gift.exception;
+package gift.product.exception;
 
 public class ProductValidationException extends RuntimeException {
